Placing the Clean Architecture configuration in another separate class for a better readability.
