+++ DancingSteps/AppDelegate.swift
@@ -15,48 +15,11 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     var window: UIWindow?
     let storyBoardName = "Main"
 
-
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
-        // Override point for customization after application launch.
-        
-        // TODO: Place this configuration in another separate class/file.
         // Configuration
-        let storyboard = UIStoryboard(name: storyBoardName, bundle: nil)
-        let tabBarController = storyboard.instantiateInitialViewController() as! UITabBarController
-        
-        // VIDEOS
-        let s_navigationController = tabBarController.viewControllers?.first as! UINavigationController
-        let s_tableViewController = s_navigationController.topViewController as! VideosTableViewController
-        let realmRepo = RealmRepo<Video>()
-        let getVideosUseCase = GetVideosUseCase(repository: realmRepo)
-        let s_presenter = VideosPresenter(useCase: getVideosUseCase)
-        getVideosUseCase.presenter = s_presenter
-        s_tableViewController.presenter = s_presenter
-        
-        // RECORDING
-        let r_navigationController = tabBarController.viewControllers?[1] as! UINavigationController
-        let recordingViewController = r_navigationController.topViewController as! CameraViewController
-        // TODO: Connect the architecture for the recording workflow.
-//        let realmRepo = RealmRepo<Video>()
-//        let saveNewVideoUseCase = SaveNewVideoUseCase(repository: realmRepo)
-//        let r_presenter = CompleteVideoPresenter(useCase: saveNewVideoUseCase)
-//        // SaveNewVideoUseCase.presenter = r_presenter
-//        s_tableViewController.presenter = s_presenter
-        
-        // TOP CHART
-        let navigationController = tabBarController.viewControllers?[2] as! UINavigationController
-        let tableViewController = navigationController.topViewController as! CongressesTableViewController
-        let useCaseNetworkProvider = UseCaseNetworkProvider()
-        let getCongressesUseCase = useCaseNetworkProvider.makeGetCongressesUseCase()
-        let presenter = CongressesPresenter(useCase: getCongressesUseCase)
-        getCongressesUseCase.presenter = presenter
-        tableViewController.presenter = presenter
-        
-        // START THE APP !
+        let mainStoryboard = UIStoryboard(name: storyBoardName, bundle: nil)
         window = UIWindow(frame: UIScreen.main.bounds)
-        window?.rootViewController = tabBarController
-        window?.makeKeyAndVisible()
-        
+        Application.shared.cleanArchitectureConfiguration(storyboard: mainStoryboard, mainWindow: window!)
         return true
     }
